project1_invoice_processing/requirements.txt
streamlit==1.32.0
langchain==0.1.13
langchain-groq==0.1.0
langchain-community==0.0.29
pdfplumber==0.11.0
pydantic==2.6.4
fpdf==1.7.2
pytesseract==0.3.10
Pillow==10.2.0
